[REFACTOR] Record 기반 쿼리 로직 삭제 #185
